Added logic to allow the user to add new employees when selecting the add an employee option. Also refactored the code by creating a connection.js file to hold the mysql2 connection and created a queries file to hold all of the db.query functions.
